+++ server/metrics.c
@@ -2,7 +2,7 @@
 
 static uint32_t historicConnections;
 static uint32_t concurrentConnections;
-static uint32_t bytesTransferred;
+static uint64_t bytesTransferred;
 
 void init_metrics() {
     historicConnections = 0;
@@ -19,7 +19,7 @@ void report_closed_connection() {
     concurrentConnections--;
 }
 
-void report_transfer_bytes(uint32_t bytes) {
+void report_transfer_bytes(uint64_t bytes) {
     bytesTransferred += bytes;
 }
 
@@ -29,6 +29,6 @@ uint32_t get_historic_connections() {
 uint32_t get_concurrent_connections() {
     return concurrentConnections;
 }
-uint32_t get_bytes_transferred() {
+uint64_t get_bytes_transferred() {
     return bytesTransferred;
 }
